pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kern"
version = "1.0.0-rc1"
description = "KERN: privacy-first local AI workspace for Windows with controlled tools, local voice output, and local memory."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "cryptography>=44.0.0",
  "httpx>=0.27.0",
  "numpy>=2.1.0",
  "sounddevice>=0.4.7",
  "pyttsx3>=2.98",
]

[project.optional-dependencies]
local_brain = [
  "llama-cpp-python>=0.3.0",
]
hf_adapter = [
  "torch>=2.6.0",
  "transformers>=4.54.0",
  "peft>=0.18.0",
  "accelerate>=1.0.0",
  "safetensors>=0.4.0",
]
documents = [
  "pymupdf>=1.24.0",
  "python-docx>=1.1.0",
  "openpyxl>=3.1.0",
]
ocr = [
  "paddleocr>=3.0.0",
  "paddlepaddle>=3.0.0",
]
scheduler = [
  "croniter>=2.0.0",
  "watchdog>=4.0.0",
]
system_control = [
  "psutil>=6.0.0",
  "pywin32>=308; platform_system == 'Windows'",
]
vector = [
  "sqlite-vec>=0.1.0",
]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.setuptools]
packages = ["app", "app.tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
